feat: OSC-34: Make tests union
